feat: table sorting and pagination
